Complete first working draft of DOVE integration
